修正8

+++ db/migrate/20250113054256_remove_email_from_users.rb
@@ -1,5 +1,8 @@
 class RemoveEmailFromUsers < ActiveRecord::Migration[7.2]
   def change
-    remove_column :users, :email, :string
+    # emailカラムが存在する場合のみ削除
+    if column_exists?(:users, :email)
+      remove_column :users, :email, :string
+    end
   end
 end
